+++ src/main/java/com/testquality/jenkins/exception/HttpException.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.testquality.jenkins;
+package com.testquality.jenkins.exception;
 
 import okhttp3.Response;
